requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==1.1.2
idna==3.6
multidict==6.0.4
numpy==1.24.4
pandas==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
twilio==8.11.0
tzdata==2023.3
urllib3==2.1.0
wheel==0.42.0
yarl==1.9.4
